move non-deprecated helpers out of the deprecated files
